==> 10x_scRNAseq_1/2_annotation_fig_1.Rmd <==
---
title: '2_annotation_fig_1'
author: "Ryuta Yokogawa"
output:
  html_document:
    number_sections: yes
    toc: true
    toc_float: true
    code_folding: show
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "~/Library/CloudStorage/OneDrive-KyotoUniversity/10xanalysis/RMD for analysis/Murase_paper/Final")
```
data: 

Hs_in-vitro (NCLCN#261　d6c11, d6c56, d6c86, d6c117)
Hs_in-vivo (AC7w, AC9w, AC10w, YK11w, AC13w, AC16w)
res1.3


# Set up

```{r}
filename<-"2_annotation_fig_1"
orig.ident<-c("c11","c56","c86","c117","AC7w","AC9w","AC10w","YK11w","AC13w","AC16w")
fig.ident<-c("c11","c56","c86","c117","7wpf","9wpf","10wpf","11wpf","13wpf","16wpf")
dataset<-c(rep("Hs_in-vitro",4),rep("Hs_in-vivo",6))
mapping<-"cr60GRCh38p12chrxmv1b1"
```

## Load libraries

```{r,message=FALSE,warning=FALSE}
suppressPackageStartupMessages({
  library(tidyverse)
  library(SeuratWrappers)
  library(patchwork)
  library(Seurat)
  library(Matrix)
  library(scales)
  library(data.table)
  library(gridExtra)
  library(knitr)
  library(scran)
  library(scater)
  library(pheatmap)
  library(gplots)
  library(RColorBrewer)
  library(scales) 
})
set.seed(0)
```

## Import data

```{r}
seurat <- readRDS("~/Library/CloudStorage/OneDrive-KyotoUniversity/10xanalysis/RMD for analysis/Murase_paper/Final/2_annotation_vivoACYK_c11_56_86_117_res0.8_5000HVGs_exp2in5cells.rds")
seurat$fig.ident<-factor(seurat$orig.ident,labels=fig.ident)
```


```{r}
theme_set(theme(text=element_text(size = 18),
                panel.background = element_rect(fill = "white",color="gray20"),
                panel.grid=element_line(color="gray92"),
                legend.key = element_rect(colour = NA, fill = NA)))
source("~/Library/CloudStorage/OneDrive-KyotoUniversity/10xanalysis/core_functions.R")
```

```{r}
celltype_pal<-c(colorRampPalette(c("white","#008000"))(4)[4:3],
                colorRampPalette(c("white","#005080"))(4)[4:3],
                colorRampPalette(c("white","red"))(4)[4:2],
                "gold",
                "#b35900",
                "#a020f0"
                ) %>% 
  setNames(c("VEM1","VEM2","EM1","EM2","M1","M2","M3","PLL1","PLL2","ZPD"))
```

# Analysis
## Annotation

```{r,fig.width=13,fig.height=5,warning=FALSE}
p1<-DimPlot(object = seurat, reduction="umap", pt.size = 0.5,group.by="cluster",label=FALSE,label.size = 5)+
  scale_color_manual(values=celltype_pal)
p2<-DimPlot(object = seurat, reduction="umap", pt.size = 0.5, group.by = "fig.ident")

(p1+p2)

pdf(paste0(filename,"_C1_1.pdf"), width = 10, height = 10)
(p1)
dev.off()

ggsave(paste0(filename,"_C1_1.png"),p1,width=10,height=10,dpi=1200)

```

```{r}
table(seurat$orig.ident,seurat$cluster) %>% 
  addmargins()
```


## Filtering
skip 

## Distribution on UMAP

```{r,fig.width=16,fig.height=8}
seurat$temp<-ifelse(seurat$fig.ident%in%c("10wpf","11wpf"),"10/11wpf",
                    ifelse(seurat$fig.ident%in%c("13wpf","16wpf"),"13/16wpf",as.character(seurat$fig.ident))) %>% 
  factor(levels=c("c11", "c56", "c86", "c117", "7wpf", "9wpf", "10/11wpf","13/16wpf"))
dat<-seurat[["umap"]][[]] %>% as.data.frame()
dat$orig.ident<-seurat$temp
plts<-levels(dat$orig.ident) %>%
  lapply(function(x){
    ggplot(NULL)+
      geom_point(dat,mapping=aes(x=UMAP_1,y=UMAP_2),color="lightgrey",size=0.5)+
      geom_point(filter(dat,orig.ident==x),mapping=aes(x=UMAP_1,y=UMAP_2),color="red",size=0.5)+
      ggtitle(x)+rd_theme()+
      theme(plot.title=element_text(size=7,hjust=0.5,vjust=1))+
      theme(axis.title=element_blank(),
                axis.text=element_blank(),
                axis.ticks=element_blank(),
                legend.position = "none")
  })

wrap_plots(plts, ncol = 4)

pdf(paste0(filename,"_D.pdf"), width = 16, height = 8)
wrap_plots(plts, ncol = 4)
dev.off()

ggsave(paste0(filename,"_D.png"),wrap_plots(plts, ncol = 4),width=20,height=10,dpi=1200)
```

## Marker expression [annotationつけた後にみる場合]

```{r}
markers <- c("POU5F1","NANOG","SOX17","TFAP2C","GTSF1","DAZL","PRAME","SYCP3","DDX4","ANHX","ZGLP1","REC8","SYCP1","FIGLA","MKI67","CCNA2") %>% intersect(rownames(seurat))
p1<-FeaturePlot(object = seurat, features = unlist(markers), min.cutoff = "q9", pt.size = 0.5, ncol=7, reduction="umap",label=FALSE)&rd_theme_italic()
ggsave(p1,filename=paste0(filename,"_markers.png"),width=25,height=11)
```

"POU5F1","NANOG","SOX17","TFAP2C","GTSF1","DAZL","PRAME","SYCP3","DDX4","ANHX","ZGLP1","REC8","SYCP1","FIGLA","MKI67","CCNA2","MCM6","PCNA","DNMT1","DNMT3A","DNMT3B","DNMT3L","UHRF1","TET1","TET2","TET3"

```{r,fig.width=6, fig.height=7}
markers<-intersect(c("POU5F1","NANOG","SOX17","TFAP2C","GTSF1","DAZL","PRAME","SYCP3","DDX4","ANHX","ZGLP1","REC8","SYCP1","FIGLA","MKI67","CCNA2"),rownames(seurat)) 
mat<-levels(seurat$cluster) %>% 
  lapply(function(y){
    if(y%in%seurat$cluster){
      if (sum(seurat$cluster==y)>1){
        seurat[["RNA"]]@data[markers,seurat$cluster==y] %>% 
          rowMeans()
        }else{
          seurat[["RNA"]]@data[markers,seurat$cluster==y]
        }
      }else{
          rep(NA,length(markers))
      }
  }) %>% do.call(cbind,.) %>% 
  as.data.frame() %>% 
  `colnames<-`(paste0(levels(seurat$cluster)))
mat_col <- data.frame(cluster=levels(seurat$cluster) %>% factor(levels(seurat$cluster)),
                      row.names = colnames(mat))
col_pal = list(cluster = celltype_pal)
breaksList = c(seq(0, 1.99, by = 0.01),seq(2, 5, by = 0.05))
newnames <- lapply(
  rownames(mat),
  function(x) bquote(bolditalic(.(x))))

pheatmap(mat,cluster_rows = F,cluster_cols = FALSE, 
             color=colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(length(breaksList)),
             breaks = breaksList,
             annotation_col=mat_col,
             annotation_colors = col_pal[1],
             show_colnames = F, show_rownames=T,
         labels_row = as.expression(newnames),
         #gaps_col = ncol(mat)-2
         )
```

```{r}
# 以前のコード...
seurat$method <- ifelse(seurat$orig.ident %in% c("c11", "c56", "c86", "c117"), "in vivo", "in vitro") %>% 
  factor(levels = c("in vitro", "in vivo"))

for (i in levels(seurat$method)){
  seu <- subset(seurat, method == i)
  mat <- levels(seu$cluster) %>% 
    lapply(function(y){
      if(y %in% seu$cluster){
        if (sum(seu$cluster == y) > 1){
          seu[["RNA"]]@data[markers, seu$cluster == y] %>% 
            rowMeans()
        } else {
          seu[["RNA"]]@data[markers, seu$cluster == y]
        }
      } else {
        rep(NA, length(markers))
      }
    }) %>% do.call(cbind, .) %>% 
    as.data.frame() %>% 
    `colnames<-`(paste0(levels(seu$cluster)))
  mat_col <- data.frame(cluster = levels(seu$cluster) %>% factor(levels(seu$cluster)),
                        row.names = colnames(mat))
  col_pal = list(cluster = celltype_pal)
  breaksList = c(seq(0, 1.99, by = 0.01),seq(2, 5, by = 0.05))
  newnames <- lapply(
    rownames(mat),
    function(x) bquote(bolditalic(.(x)))
  )

  # PDF出力用にファイルを開く
  pdf(paste0(filename,"_sup_A_",i,".pdf"), width = 4, height = 4)
  
  # ヒートマップの描画
  pheatmap(mat, cluster_rows = F, cluster_cols = FALSE, 
           color = colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(length(breaksList)),
           breaks = breaksList,
           annotation_col = mat_col,
           annotation_colors = col_pal[1],
           show_colnames = F, show_rownames = T,
           labels_row = as.expression(newnames),
           main = i)
  
  # PDF出力を終了
  dev.off()
}
```


## Marker expression [vivo vs vitro]

```{r,fig.width=4, fig.height=4.5}
seurat$method<-ifelse(seurat$orig.ident%in%c("c11","c56","c86","c117"),"in vitro","in vivo") %>% 
  factor(levels=c("in vitro","in vivo"))
for (i in levels(seurat$method)){
  seu<-subset(seurat,method==i)
  mat<-levels(seu$cluster) %>% 
  lapply(function(y){
    if(y%in%seu$cluster){
      if (sum(seu$cluster==y)>1){
        seu[["RNA"]]@data[markers,seu$cluster==y] %>% 
          rowMeans()
        }else{
          seu[["RNA"]]@data[markers,seu$cluster==y]
        }
      }else{
          rep(NA,length(markers))
      }
  }) %>% do.call(cbind,.) %>% 
  as.data.frame() %>% 
  `colnames<-`(paste0(levels(seu$cluster)))
mat_col <- data.frame(cluster=levels(seu$cluster) %>% factor(levels(seu$cluster)),
                      row.names = colnames(mat))
col_pal = list(cluster = celltype_pal)
breaksList = c(seq(0, 1.99, by = 0.01),seq(2, 5, by = 0.05))
newnames <- lapply(
  rownames(mat),
  function(x) bquote(bolditalic(.(x))))

pheatmap(mat,cluster_rows = F,cluster_cols = FALSE, 
             color=colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(length(breaksList)),
             breaks = breaksList,
             annotation_col=mat_col,
             annotation_colors = col_pal[1],
             show_colnames = F, show_rownames=T,
         labels_row = as.expression(newnames),
         #gaps_col = ncol(mat)-2,
         main=i)

pdf(paste0(filename,"_sup_A_",i,".pdf"), width = 4, height = 4)
pheatmap(mat,cluster_rows = F,cluster_cols = FALSE, 
             color=colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(length(breaksList)),
             breaks = breaksList,
             annotation_col=mat_col,
             annotation_colors = col_pal[1],
             show_colnames = F, show_rownames=T,
         labels_row = as.expression(newnames),
         #gaps_col = ncol(mat)-2,
         main=i)
dev.off()
}
```

```{r}
# Update the method field based on 'orig.ident'
seurat$method <- ifelse(seurat$orig.ident %in% c("c11", "c56", "c86", "c117"), "in vitro", "in vivo") %>% 
  factor(levels = c("in vitro", "in vivo"))

# Function to calculate mean expression and save to CSV
calculateAndSaveExpression <- function(seurat_object, condition, filename) {
  seu <- subset(seurat_object, method == condition)
  
  if (nrow(seu) > 0) { # Check if the subset is not empty
    mat <- levels(seu$cluster) %>% 
      lapply(function(y) {
        if(y %in% seu$cluster) {
          if (sum(seu$cluster == y) > 1) {
            seu[["RNA"]]@data[markers, seu$cluster == y] %>% rowMeans()
          } else {
            seu[["RNA"]]@data[markers, seu$cluster == y]
          }
        } else {
          rep(NA, length(markers))
        }
      }) %>% do.call(cbind, .) %>% as.data.frame() %>% `colnames<-`(paste0(levels(seu$cluster)))
    
    # Save to CSV
    write.csv(mat, paste0(filename, "_", condition, ".csv"))
  }
}

# Calculate and save for combined, vivo, and vitro conditions
calculateAndSaveExpression(seurat, "in vivo", "expression_levels_vivo")
calculateAndSaveExpression(seurat, "in vitro", "expression_levels_vitro")

```




```{r}
dat<-seurat[["RNA"]]@data[markers,] %>% 
  as.matrix() %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var="cells") %>% 
  tidyr::pivot_longer(cols=-cells,names_to="genes",values_to="exp") %>% 
  transform(genes=factor(genes,levels=markers))
dat$cluster<-seurat@meta.data[dat$cells,"cluster"] %>% 
  factor(levels=levels(.) )
```

```{r,fig.width=14,fig.height=20}
p1<-ggplot(dat,aes(x=cluster,y=exp))+
  geom_violin(aes(fill=cluster),scale="width")+
  xlab("")+
  ylab("")+
  facet_wrap(genes~.,ncol=1, strip.position = "right")+
  theme_bw()+
  theme(panel.background = element_blank(),
        panel.grid = element_line(color = "white"), 
        strip.text.y = element_text(angle = 0,size=18, face="bold",hjust=0),
        axis.text.x = element_text(size=18,angle=45,hjust=1, face="bold"),
        strip.placement = "outside",
        strip.background = element_blank(),
        legend.position="none")+
  scale_fill_manual(values=celltype_pal)
p1
```

## Population of germ cells in each stage

```{r,fig.width=2.5,fig.height=2}
temp<-seurat@meta.data %>% 
  transform(cluster=factor(cluster,levels=levels(cluster))) %>% 
  group_by(fig.ident,cluster) %>% 
  summarise(count=n()) %>% 
  mutate(ratio=(count/sum(count)*100) %>% 
           round(digits=1) %>% 
           format(nsmall=1)) %>% 
  arrange(fig.ident, desc(cluster)) %>% 
  mutate(label_y = (cumsum(count) - 0.5 * count) / sum(count)) 

p1<-ggplot(temp,aes(x=fig.ident,y=count,fill=cluster))+
  geom_bar(stat="identity",position="fill",color="black")+
  scale_y_continuous(labels=percent)+
  theme_bw(base_size = 6)+
  ylab("Percentage of germ cells")+
  xlab("")+
  scale_fill_manual(values=celltype_pal)+
  theme_classic(base_size = 7)+
  theme(axis.text.x = element_text(size=7,angle=90,hjust=1,vjust=0.5, color="black", face ="bold"),
        axis.text.y = element_text(size=7),
        axis.title = element_text(size=7,face="bold"),
        axis.line.x = element_line(size=0.5, color="black"),
        axis.line.y = element_line(size=0.5, color="black"),
        legend.text=element_text(size=7, color="black"),
        legend.title=element_blank(),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), panel.grid.major.y = element_blank(),
        panel.background = element_rect(fill='white'),
        strip.background = element_blank(),
        strip.text.x = element_blank()
)

p1

pdf(paste0(filename,"_G.pdf"), width = 2.5, height = 2)
p1
dev.off()
```

```{r}
temp<-seurat@meta.data %>% 
  transform(cluster=factor(cluster,levels=levels(cluster))) %>% 
  group_by(fig.ident,cluster) %>% 
  summarise(count=n()) %>% 
  mutate(ratio=(count/sum(count)*100) %>% 
           round(digits=1) %>% 
           format(nsmall=1)) %>% 
  arrange(fig.ident, desc(cluster)) %>% 
  mutate(label_y = (cumsum(count) - 0.5 * count) / sum(count)) 

p1<-ggplot(temp,aes(x=fig.ident,y=count,fill=cluster))+
  geom_bar(stat="identity",position="fill",color="black")+
  geom_text(aes(label=ratio, y=label_y), position=position_fill(vjust=0.5), size=2.5, color="black") +
  scale_y_continuous(labels=percent)+
  theme_bw(base_size = 6)+
  ylab("Percentage of germ cells")+
  xlab("")+
  scale_fill_manual(values=celltype_pal)+
  theme_classic(base_size = 7)+
  theme(axis.text.x = element_text(size=7,angle=90,hjust=1,vjust=0.5, color="black", face ="bold"),
        axis.text.y = element_text(size=7),
        axis.title = element_text(size=7,face="bold"),
        axis.line.x = element_line(size=0.5, color="black"),
        axis.line.y = element_line(size=0.5, color="black"),
        legend.text=element_text(size=7, color="black"),
        legend.title=element_blank(),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), panel.grid.major.y = element_blank(),
        panel.background = element_rect(fill='white'),
        strip.background = element_blank(),
        strip.text.x = element_blank()
)

p1

pdf(paste0(filename,"_Gpercent.pdf"), width = 5, height = 5)
p1
dev.off()

```


```{r,fig.width=5,fig.height=5}
ggplot(temp,aes(x=fig.ident,y=count,fill=cluster))+
  geom_bar(stat="identity",position="fill",color="black")+
  geom_text(aes(label = count, y = label_y),size=3,hjust=0.5)+
  scale_y_continuous(labels=percent)+
  theme_bw(base_size = 15)+
  ylab("Percentage of germ cells")+
  xlab("")+
  scale_fill_manual(values=celltype_pal)+
  theme_classic(base_size = 15)+
  theme(axis.text.x = element_text(size=18,angle=45,hjust=1, color="black", face ="bold"),
        axis.text.y = element_text(size=15),
        axis.title = element_text(size=18,face="bold"),
        axis.line.x = element_line(size=0.5, color="black"),
        axis.line.y = element_line(size=0.5, color="black"),
        legend.text=element_text(size=15, color="black"),
        legend.title=element_blank(),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), panel.grid.major.y = element_blank(),
        panel.background = element_rect(fill='white'),
        strip.background = element_blank(),
        strip.text.x = element_blank()
)

# 既存のggplotコード
p <- ggplot(temp, aes(x=fig.ident, y=count, fill=cluster)) +
  geom_bar(stat="identity", position="fill", color="black") +
  geom_text(aes(label = count, y = label_y), size=3, hjust=0.5) +
  scale_y_continuous(labels=percent) +
  theme_bw(base_size = 15) +
  ylab("Percentage of germ cells") +
  xlab("") +
  scale_fill_manual(values=celltype_pal) +
  theme_classic(base_size = 15) +
  theme(axis.text.x = element_text(size=18, angle=45, hjust=1, color="black", face ="bold"),
        axis.text.y = element_text(size=15),
        axis.title = element_text(size=18, face="bold"),
        axis.line.x = element_line(size=0.5, color="black"),
        axis.line.y = element_line(size=0.5, color="black"),
        legend.text = element_text(size=15, color="black"),
        legend.title = element_blank(),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), panel.grid.major.y = element_blank(),
        panel.background = element_rect(fill='white'),
        strip.background = element_blank(),
        strip.text.x = element_blank())

# PDFとして保存
ggsave("2_annotation_fig_1_Gnum.pdf", plot=p, device="pdf", width=5, height=5)
```

## Population of germ cells in each cluster

```{r}
fig_pal<-c("#fce5cd","#f9cb9c","#f6b26b","#e69138","#cfe2f3","#9fc5e8","#6fa8dc","#3d85c6","#0b5394","#073763") %>% 
  setNames(c("c11","c56","c86","c117","7wpf","9wpf","10wpf","11wpf","13wpf","16wpf"))
```


```{r,fig.width=2.5,fig.height=2}
temp<-seurat@meta.data %>% 
  transform(cluster=factor(cluster,levels=levels(cluster))) %>% 
  group_by(fig.ident,cluster) %>% 
  summarise(count=n()) %>% 
  mutate(ratio=(count/sum(count)*100) %>% 
           round(digits=1) %>% 
           format(nsmall=1)) %>% 
  arrange(fig.ident, desc(fig.ident)) %>% 
  mutate(label_y = (cumsum(count) - 0.5 * count) / sum(count))

p1<-ggplot(temp,aes(x=cluster,y=count,fill=fig.ident))+
  geom_bar(stat="identity",position="fill",color="black")+
  scale_y_continuous(labels=percent)+
  theme_bw(base_size = 6)+
  ylab("Percentage of germ cells")+
  xlab("")+
  scale_fill_manual(values=fig_pal)+
  theme_classic(base_size = 7)+
  theme(axis.text.x = element_text(size=7,angle=90,hjust=1,vjust=0.5, color="black", face ="bold"),
        axis.text.y = element_text(size=7),
        axis.title = element_text(size=7,face="bold"),
        axis.line.x = element_line(size=0.5, color="black"),
        axis.line.y = element_line(size=0.5, color="black"),
        legend.text=element_text(size=7, color="black"),
        legend.title=element_blank(),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), panel.grid.major.y = element_blank(),
        panel.background = element_rect(fill='white'),
        strip.background = element_blank(),
        strip.text.x = element_blank()
)

p1
pdf(paste0(filename,"_H.pdf"), width = 2.8, height = 2)
p1
dev.off()
```

```{r,fig.width=5,fig.height=5}
temp <- seurat@meta.data %>% 
  transform(cluster=factor(cluster,levels=levels(cluster))) %>% 
  group_by(fig.ident,cluster) %>% 
  summarise(count=n()) %>% 
  mutate(ratio=(count/sum(count)*100) %>% 
           round(digits=1) %>% 
           format(nsmall=1)) %>% 
  arrange(fig.ident, desc(fig.ident)) %>% 
  mutate(label_y = (cumsum(count) - 0.5 * count) / sum(count))

p1 <- ggplot(temp,aes(x=cluster,y=count,fill=fig.ident))+
  geom_bar(stat="identity",position="fill",color="black")+
  scale_y_continuous(labels=percent)+
  theme_bw(base_size = 6)+
  ylab("Percentage of germ cells")+
  xlab("")+
  scale_fill_manual(values=fig_pal)+
  theme_classic(base_size = 7)+
  theme(axis.text.x = element_text(size=7,angle=90,hjust=1,vjust=0.5, color="black", face ="bold"),
        axis.text.y = element_text(size=7),
        axis.title = element_text(size=7,face="bold"),
        axis.line.x = element_line(size=0.5, color="black"),
        axis.line.y = element_line(size=0.5, color="black"),
        legend.text=element_text(size=7, color="black"),
        legend.title=element_blank(),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), panel.grid.major.y = element_blank(),
        panel.background = element_rect(fill='white'),
        strip.background = element_blank(),
        strip.text.x = element_blank()) +
  geom_text(aes(label = count, y = label_y), size = 2.5, position = position_fill(vjust = 0.5))

p1
pdf(paste0(filename,"_Hnumber.pdf"), width = 5, height = 5)
p1
dev.off()

```



## Cell cycle (QuickGO meiotic cell cycle)

```{r}
cellcycle_pal<-c("#1F77B4","#FF7F0E","#2CA02C",
                "red") %>% 
  setNames(.,c("G1","S","G2M","Meiosis"))
```

```{r}
S<-cc.genes.updated.2019$s.genes
G2M<-cc.genes.updated.2019$g2m.genes
Meiosis<-fread("~/Library/CloudStorage/OneDrive-KyotoUniversity/10xanalysis/Template/QuickGO_meiotic cell cycle.tsv",data.table = F)
Meiosis<-unique(Meiosis$SYMBOL)
cell_cycle<-data.frame(gene=c(S,G2M,Meiosis),
                       stage=c(rep("S",length(S)),rep("G2M",length(G2M)),rep("Meiosis",length(Meiosis))))
fun<-function(x){
  unique(x) %>% 
    length()
  }

a<-cell_cycle[cell_cycle$gene%in%cell_cycle$gene[duplicated(cell_cycle$gene)],] %>% 
  arrange(gene) %>% 
  group_by(gene) %>% 
  summarise(freq=fun(stage)) %>% 
  filter(freq>1)
a

cell_cycle<-filter(cell_cycle,duplicated(gene)==F) %>% 
  filter(gene%in%a$gene==FALSE)
```


```{r}
expressed_gene<-rownames(seurat)[rowSums(seurat[["RNA"]]@counts>=1)>2]
cellcycle<-filter(cell_cycle,gene%in%expressed_gene)
## cellcycle$gene<-fromHS(cellcycle$gene,"Mf",HsMf=TRUE)
table(cellcycle$stage)
```

```{r,fig.width=7,fig.height=7}
for (x in unique(cellcycle$stage)){
  gene<-cellcycle[cellcycle$stage==x,"gene"]
  mat<-seurat[["RNA"]]@data[gene,seurat$method=="in vivo"] %>% 
    as.matrix() %>% 
    t()
  mat<-cbind(mat,ref=rowMeans(mat))
  pheatmap(cor(mat),main=x)
} 
```

```{r}
CellCycleScoring. <- function(
  object,
  s.features,
  g2m.features,
  meiosis.features,
  ctrl = NULL,
  set.ident = FALSE,
  ...
) {
  name <- 'Cell.Cycle'
  features <- list('S' = s.features, 'G2M' = g2m.features, 'Meiosis' = meiosis.features)
  if (is.null(x = ctrl)) {
    ctrl <- min(vapply(X = features, FUN = length, FUN.VALUE = numeric(length = 1)))
  }
  object.cc <- AddModuleScore(
    object = object,
    features = features,
    name = name,
    ctrl = ctrl,
    ...
  )
  cc.columns <- grep(pattern = name, x = colnames(x = object.cc[[]]), value = TRUE)
  cc.scores <- object.cc[[cc.columns]]
  rm(object.cc)
  # CheckGC()
  assignments <- apply(
    X = cc.scores,
    MARGIN = 1,
    FUN = function(scores, null = 'G1') {
      if (all(scores < 0)) {
        return(null)
      } else {
        if (length(which(x = scores == max(scores))) > 1) {
          return('Undecided')
        } else {
          return(names(features)[which(x = scores == max(scores))])
        }
      }
    }
  )
  cc.scores <- merge(x = cc.scores, y = data.frame(assignments), by = 0)
  colnames(x = cc.scores) <- c('rownames', paste0(names(features),".Score"), 'Phase')
  rownames(x = cc.scores) <- cc.scores$rownames
  cc.scores <- cc.scores[, c(paste0(names(features),".Score"), 'Phase')]
  object[[colnames(x = cc.scores)]] <- cc.scores
  if (set.ident) {
    object[['old.ident']] <- Idents(object = object)
    Idents(object = object) <- 'Phase'
  }
  return(object)
}
```

```{r}
filtered_cellcycle<-cellcycle
seurat <- CellCycleScoring.(seurat, 
                            s.features = filtered_cellcycle$gene[filtered_cellcycle$stage%in%c("S")],
                            g2m.features = filtered_cellcycle$gene[filtered_cellcycle$stage%in%c("G2M")],
                            meiosis.features = filtered_cellcycle$gene[filtered_cellcycle$stage%in%c("Meiosis")])
seurat$Phase<-factor(seurat$Phase,levels=c("G1","S","G2M","Meiosis"))
```


```{r}
RidgePlot(seurat, features = c("PCNA", "TOP2A", "MCM6", "MKI67"), ncol = 2, group.by="Phase")&
  scale_fill_manual(values=cellcycle_pal)
```

```{r,fig.width=6.5 ,fig.height=5}
pdat<-seurat[["umap"]]@cell.embeddings %>% 
  as.data.frame() %>% 
  cbind(seurat@meta.data)
pdat<-pdat[order(seurat$cluster),]
p1<-ggplot(pdat,aes(x=UMAP_1,y=UMAP_2,color=Phase))+
  geom_point(size=0.5)+
  scale_color_manual(values=cellcycle_pal)+
  rd_theme(legend.position = "right")+
  guides(color = guide_legend(override.aes = list(size = 5)))+
  theme(axis.line = element_line(colour = 'black', size = 1))
p1

pdf(paste0(filename,"_C2_1.pdf"), width = 10, height = 10)
(p1)
dev.off()

ggsave(paste0(filename,"_C2_1.png"),p1,width=10,height=10,dpi=1200)
```

```{r,fig.width=8,fig.height=8}
pdat<-seurat[["umap"]]@cell.embeddings %>% 
  as.data.frame() %>% 
  cbind(seurat@meta.data)
pdat<-pdat[order(seurat$cluster),]
p1<-ggplot(pdat,aes(x=UMAP_1,y=UMAP_2,color=Phase))+
  geom_point(size=1.5,show.legend = F)+
  scale_color_manual(values=cellcycle_pal)+
  rd_theme(legend.position = "right")+
  guides(color = guide_legend(override.aes = list(size = 5)))+
  theme(axis.line = element_line(colour = 'black', size = 1))
p1

# pdf(paste0(filename,"_rough_Tirosh_cellcycle_UMAP_simple.pdf"), width = 8, height = 8)
# print(p1)#メインのグラフ
# dev.off()
```

### Population of germ cells in each stage

```{r}
mytheme<-function(){
  theme_minimal()+
  theme(axis.text.x = element_text(size=25, color="black",face="bold",angle=90,hjust=1,vjust=0.5),
        axis.text.y = element_text(size=18, color="black",face="bold"),
        legend.text=element_text(size=17, color="black"),
        legend.title=element_text(size=19, color="black"),
        axis.line.x = element_line(size=1.2, color="black"),
        axis.line.y = element_line(size=1.2, color="black"),
        strip.text = element_text(size=30, face="bold",color="black"),
        axis.title = element_text(size=22,face="bold"),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.major.y = element_blank(),
        panel.background = element_blank(),
        axis.ticks=element_line(size=0.8),
        axis.ticks.length.y=unit(0.2,units="cm"),
        strip.background = element_blank()
        )
}
```

```{r,fig.width=8,fig.height=5.5}
seurat$sample<-seurat$orig.ident
temp<-seurat@meta.data %>% 
  group_by(sample,Phase) %>% 
  summarise(count=n()) %>% 
  mutate(ratio=(count/sum(count)*100) %>% 
           round(digits=1) %>% 
           format(nsmall=1)) %>% 
  arrange(sample, desc(Phase)) %>% 
  mutate(label_y = (cumsum(count) - 0.5 * count) / sum(count))

```

```{r,fig.width=6,fig.height=5}
p1<-ggplot(temp,aes(x=sample,y=count,fill=Phase))+
  geom_bar(stat="identity",position="fill",color="black",size=0.8)+
  #geom_text(aes(label = count, y = label_y),size=3,hjust=0.5)+
  scale_y_continuous(labels=percent)+
  theme_bw(base_size = 15)+
  ylab("Percentage of germ cells")+
  xlab("")+
  scale_fill_manual(values=cellcycle_pal)+
  mytheme()
p1
# pdf(paste0(filename,"_rough_Tirosh_cellcycle_simple.pdf"),width=6, height=5)
# p1
# dev.off()
```

```{r,fig.width=8,fig.height=5.5}
seurat$sample<-seurat$cluster
temp<-seurat@meta.data %>% 
  group_by(sample,Phase) %>% 
  summarise(count=n()) %>% 
  mutate(ratio=(count/sum(count)*100) %>% 
           round(digits=1) %>% 
           format(nsmall=1)) %>% 
  arrange(sample, desc(Phase)) %>% 
  mutate(label_y = (cumsum(count) - 0.5 * count) / sum(count))
```

```{r,fig.width=8,fig.height=5.5}
ggplot(temp,aes(x=sample,y=count,fill=Phase))+
  geom_bar(stat="identity",position="fill",color="black")+
  geom_text(aes(label = count, y = label_y),size=3,hjust=0.5)+
  scale_y_continuous(labels=percent)+
  theme_bw(base_size = 15)+
  ylab("Percentage of germ cells")+
  xlab("")+
  scale_fill_manual(values=cellcycle_pal)+
  mytheme()
```

## Key genes{.tabset .tabset-fade}
### figure

```{r,echo=FALSE,fig.width = 15, fig.height = 30, fig.margin=TRUE}
rd_theme_italic2<-function(legend.position="none"){theme(axis.title=element_blank(),
                                  axis.text=element_blank(),
                                  axis.ticks=element_blank(),
                                  panel.grid=element_blank(),
                                  legend.position = legend.position,
                                  panel.background = element_blank(),
                                  panel.border = element_blank(),
                                  axis.line = element_line(size=0.5),#element_blank(),
                                  plot.title= element_text(face="bold.italic",size=20)
)}

markers<-c("POU5F1","NANOG","SOX17","TFAP2C","GTSF1","DAZL","PRAME","SYCP3","DDX4",
           "ANHX","ZGLP1","REC8","SYCP1","FIGLA","MKI67","CCNA2")
p1<-FeaturePlot(object = seurat, features = markers, min.cutoff = "q9", pt.size = 0.5, ncol=3, reduction="umap",label=FALSE)&rd_theme_italic2(legend.position = "right")
p1
pdf(paste0(filename,"_F'.pdf"), width = 15, height = 30)
p1
dev.off()

ggsave(p1,filename=paste0(filename,"_F'.png"),width=15,height=30)
```

### Pluripotency

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("POU5F1","NANOG","SOX2","LIN28A"), pt.size=1.0)
#Li, Qiao, 2017, Cell Stem Cell#
```

### Inner cell mass

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("ZFP42","TFCP2L1","KLF4"), pt.size=1.0)
```

### Early PGC

```{r,echo=FALSE,fig.width = 15, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("PRDM1","TFAP2C","NANOS3","SOX17","SOX15","SALL4","PDPN","KIT","WNT3"), pt.size=1.0)
#Fernandes, Lopes, 2018, Molecular Human Reproduction#
#Li, Qiao, 2017, Cell Stem Cell#
```

### Late PGC

```{r,echo=FALSE,fig.width = 10, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("DPPA3","DAZL","DDX4","TDRD6","PIWIL2","PRAME"), pt.size=1.0)
#Li, Qiao, 2017, Cell Stem Cell#
```

### RA-responsive

```{r,echo=FALSE,fig.width = 10, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("ZGLP1","STRA8","REC8","ANHX","ASB9","THRA"), pt.size=1.0)
#Li, Qiao, 2017, Cell Stem Cell#
```

### Synaptonemal complex

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("SYCP2","SYCP3","SYCE1","HORMAD1"), pt.size=1.0)
#Motani, Saitou, 2022#
```

### Meiotic recombination

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("DMC1","SPO11","PRDM9","HORMAD1"), pt.size=1.0)
#Motani, Saitou, 2022#
```

### Oogenic

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("FIGLA","NOBOX","ZP3"), pt.size=1.0)
#Motani, Saitou, 2022#
```


## Key pathway genes{.tabset .tabset-fade}
### WNT

```{r,echo=FALSE,fig.width = 20, fig.height = 25, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("LEF1","FZD1","FZD2","FZD3","FZD4","FZD5","FZD6","FZD7", "FZD8","FZD9","FZD10","LRP5","LRP6","ROR2","RYK","RSPO1","LGR4","LGR6","ZNRF3","RNF43"), pt.size=1.5)
```

### BMP

```{r,echo=FALSE,fig.width = 20, fig.height = 20, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("GATA2","GATA3","SMAD1","SMAD5","SMAD2","SMAD3","SMAD4","SMAD6","SMAD9","ID1","ID2","ID3","ID4"), pt.size=1.5)
```

### NOTCH

```{r,echo=FALSE,fig.width = 20, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("NOTCH1","NOTCH2","NOTCH3","HES1","HES2","HEY1","HEY2","JAG1","JAG2","DLL1","DLL3"), pt.size=1.5)
#Fernandes, Lopes, 2018, Molecular Human Reproduction#
#Li, Qiao, 2017, Cell Stem Cell#
```

### SCF

```{r,echo=FALSE,fig.width = 10, fig.height = 5, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("KIT","KITLG"), pt.size=1.5)
#Li, Qiao, 2017, Cell Stem Cell#
```

### Hedgehog

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("GLI1","GLI2","GLI3","PTCH1","GAS1","CDON","BOC","SMO"), pt.size=1.5)
```

### RA1

```{r,echo=FALSE,fig.width = 10, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("RARA","RARB","RARG","RXRA","RXRB","RXRG"), pt.size=1.5)
```

### RA2

```{r,echo=FALSE,fig.width = 10, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("NCOR1","NCOR2","ALDH1A1","ALDH1A2","ALDH1A3"), pt.size=1.5)
```

### RA3

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("FGF8","CDX1","CDX2","CDX4"), pt.size=1.5)
```

### HOXA

```{r,echo=FALSE,fig.width = 20, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("HOXA1","HOXA2","HOXA3","HOXA4","HOXA5","HOXA6","HOXA7","HOXA9","HOXA10","HOXA11","HOXA13"), pt.size=1.5)
```

### HOXB

```{r,echo=FALSE,fig.width = 20, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("HOXB1","HOXB2","HOXB3","HOXB4","HOXB5","HOXB6","HOXB7","HOXB8","HOXB9","HOXB13"), pt.size=1.5)
```

### HOXC

```{r,echo=FALSE,fig.width = 15, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("HOXC4","HOXC5","HOXC6","HOXC8","HOXC9","HOXC10","HOXC11","HOXC12","HOXC13"), pt.size=1.5)
```

### HOXD

```{r,echo=FALSE,fig.width = 15, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("HOXD4","HOXD5","HOXD6","HOXD8","HOXD9","HOXD10","HOXD11","HOXD12","HOXD13"), pt.size=1.5)
```

### HOX anterior

```{r,echo=FALSE,fig.width = 15, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("HOXA2","HOXA3","HOXA5","HOXB4","HOXB5","HOXD4","HOXD8"), pt.size=1.5)
```

### HOX posterior

```{r,echo=FALSE,fig.width = 20, fig.height = 25, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("HOXA6","HOXA9","HOXA10","HOXB6","HOXB8","HOXB9","HOXC4","HOXC5","HOXC6","HOXC8","HOXC9","HOXC10","HOXC11","HOXC12","HOXC13","HOXD10","HOXD11","HOXD13"), pt.size=1.5)
```


## Key pathway receptors{.tabset .tabset-fade}
### BMP2

```{r,echo=FALSE,fig.width = 20, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("ACVR1","ACVR2A","ACVR2B","BMPR1A","BMPR1B","BMPR2","SMO","GPC1","HJV","ENG"), pt.size=1.5)
```

### FGF2

```{r,echo=FALSE,fig.width = 20, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("FGFRL1","FGFR1","FGFR2","FGFR3","FGFR4","SDC1","SDC2","SDC3","SDC4","GPC4","NRP1"), pt.size=1.5)
```

### WNT4

```{r,echo=FALSE,fig.width = 10, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("FZD1","FZD2","FZD6","FZD8","SMO","NOTCH1"), pt.size=1.5)
```

### FGF9

```{r,echo=FALSE,fig.width = 10, fig.height = 10, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("FGFR1","FGFR2","FGFR3","FGFR4"), pt.size=1.5)
```

### SHH

```{r,echo=FALSE,fig.width = 20, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("GPC1","GPC5","GAS1","BOC","PTCH1","PTCH2","SCUBE2","SMO","HHIP","LRP2","CDON"), pt.size=1.5)
```

## From paper https://orcid.org/0000-0002-6487-1329 {.tabset .tabset-fade}
### Mitotic PGC
```{r,echo=FALSE,fig.width = 15, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("NANOG","POUF5F1","SOX13","PRDM1","KLF4","SALL4","LIN28A","TFAP2C","TFCP2L1"), pt.size=1.5)
```

### RA-responsive FGC
```{r,echo=FALSE,fig.width = 15, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("ID3","RARB","ZGLP1","HES6","ZNF362","ZBTB11","HOXA5","HOXB6","HMGB3","PBX1","SOX30","GATA3"), pt.size=1.5)
```

### Pro-meiotic FGC
```{r,echo=FALSE,fig.width = 15, fig.height = 15, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("PRDM9","LHX8","NR4A2","ZNF382","ZNF208","ZNF791","MGA","ZIC4","HSF2","DMRTA2","L3MBTL1","SOX30","GATA3","FOXD1"), pt.size=1.5)
```

## ER gene{.tabset .tabset-fade}
### early upregulated gene
```{r,echo=FALSE,fig.width = 20, fig.height = 40, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("IL22RA1", "BRDT", "LOC102723543", "DAZL", "IL12B", "KHDC1L", 
"STK31", "CHCHD2", "ZNF736", "LRRC61", "TUSC1", "C9orf64", "PLAU", 
"A2M", "LOC105369746", "GTSF1", "PIWIL1", "MEG3", "SPESP1", "SPATA22", 
"USP6", "TDRD12", "ZNF534", "NLRP9", "ZNF667", "ZNF667-AS1", 
"LOC107985433", "PRAME", "MAP7D2", "MAGEB2", "MAGEB1", "LDOC1"), pt.size=1.5)
```

### late upregulated gene
```{r,echo=FALSE,fig.width = 20, fig.height = 40, fig.margin=TRUE}
FeaturePlot(seurat,reduction = "umap",
                  features=c("LINC01160", "RHEX", "GCSAML-AS1", "EMX1", "C2CD6", "KRBOX1", 
"DCAF4L1", "PDCL2", "RBM46", "DDX4", "C5orf47", "HDGFL1", "OOEP", 
"DDX43", "LOC101928688", "LOC101929128", "PIWIL2", "LOC101926892", 
"DCAF4L2", "KCNV2", "ANKRD2", "CYP2E1", "CTSF", "GDPD4", "LOC105369655", 
"SYCP3", "ANHX", "TDRD9", "LOC105376917", "ZNF492", "DSCR8", 
"CXorf67", "LUZP4", "MAGEC1", "MAGEA11", "MAGEA4"), pt.size=1.5)
```

## Distribution on UMAP vivo vs vitro

```{r,fig.width=50,fig.height=10}
dat<-seurat[["umap"]][[]] %>% as.data.frame()
dat$orig.ident<-seurat@meta.data$method
plts<-levels(dat$orig.ident) %>%
  lapply(function(x){
    ggplot(NULL)+
      geom_point(dat,mapping=aes(x=UMAP_1,y=UMAP_2),color="lightgrey",size=1)+
      geom_point(filter(dat,orig.ident==x),mapping=aes(x=UMAP_1,y=UMAP_2),color="blue",size=1)+
      ggtitle(x)+
      theme(plot.title=element_text(size=30,hjust=0.5,vjust=1))+
      theme(axis.title=element_blank(),
                axis.text=element_blank(),
                axis.ticks=element_blank(),
                legend.position = "none")
  })

wrap_plots(plts, ncol = 5)&rd_theme()
```

```{r}

```
